# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is an example log-rush server",
        "title": "log-rush simple server",
        "contact": {
            "name": "API Support",
            "email": "contact@fabiankachlock.dev"
        },
        "license": {
            "name": "n.a.",
            "url": "n.a."
        },
        "version": "1.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
        "/batch": {
            "post": {
                "description": "add a bunch of logs at once to a stream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "log"
                ],
                "summary": "push multiple logs at once",
                "operationId": "batch",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/log": {
            "post": {
                "description": "add a new log on a logstream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "log"
                ],
                "summary": "push a log",
                "operationId": "log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stream/register": {
            "post": {
                "description": "create a new logstream on the server so that client can subscribe to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logstream"
                ],
                "summary": "register a logstream",
                "operationId": "register-stream",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.LogStream"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stream/unregister": {
            "post": {
                "description": "delete a logstream an close all open connection to it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logstream"
                ],
                "summary": "unregister a logstream",
                "operationId": "unregister-stream",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/streams": {
            "get": {
                "description": "get all available logstreams on the server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logstream"
                ],
                "summary": "list all streams",
                "operationId": "list-stream",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.LogStreamsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.LogStream": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github.com_fabiankachlock_log-rush-simple-server_log_delivery_http.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github.com_fabiankachlock_log-rush-simple-server_logstream_delivery_http.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.LogStreamsResponse": {
            "type": "object",
            "properties": {
                "streams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LogStream"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "description": "all endpoints for logstreams",
            "name": "logstream"
        },
        {
            "description": "all endpoints for logs",
            "name": "log"
        }
    ]
}